Move the contents of the `std` module to the top level.
Make this crate be "just" an implementation of std, instead of containing
an implementation of std.

Also, tighten up usage of features in dependencies, and add comments.
